+++ tests/files/src/0477_array_negation.php
@@ -1,0 +1,13 @@
+<?php
+
+function test_array_negation(array $example) {
+    $v = rand() % 2 ? ['key' => 'value'] : null;
+    if (!$v)  {
+        echo strlen($v);
+    }
+
+    $v2 = rand() % 2 ? $example : null;
+    if (!$v2)  {
+        echo strlen($v2);
+    }
+}